mejora almacenamiento

+++ include/util.h
@@ -7,7 +7,7 @@ u8 CreateInvisibleSpriteWithCallback(void (*)(struct Sprite *));
 void StoreWordInTwoHalfwords(u16 *, u32);
 void LoadWordFromTwoHalfwords(u16 *, u32 *);
 int CountTrailingZeroBits(u32 value);
-void BlendPalette(u16 palOffset, u16 numEntries, u8 coeff, u32 blendColor);
+void BlendPalette(u32 palOffset, u32 numEntries, u32 coeff, u32 blendColor);
 void DesplazaTonoPaleta(u32 offsetPaleta, u32 personalidad);
 void DesplazaTonoPaletaBuffer(u16 *buffer, u32 personalidad);
